# openapi.yaml
openapi: 3.0.0
info:
  title: API
  description: ""
  version: 1.0.0
servers:
- url: http://localhost:3000
- url: ""
paths:
...
  /api/user/signup:
    post:
      tags:
      - User Tests
      summary: Make User
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/user_signup_body"
            example:
              email: testemail2@mail.com
              username: testuser2
              password: "123456"
      responses:
        "200":
          description: succesful
...
components:
  schemas:
...
    user_signup_body:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer